moved the schema from latestpost component to the page
